requirements.txt
streamlit>=1.28.0
python-dotenv>=1.0.0
langchain>=0.0.340
langchain-community>=0.0.14
langchain-core>=0.1.16
langchain-huggingface>=0.0.2
faiss-cpu>=1.7.4  # or faiss-gpu if you have CUDA
pypdf>=3.15.0
huggingface-hub>=0.19.0
sentence-transformers>=2.2.2


fastapi>=0.115.0
uvicorn[standard]>=0.30.6
langchain-openai>=0.2.0
langchain-groq>=0.2.0
qdrant-client>=1.11.0
pydantic>=2.9.0
python-multipart>=0.0.12
chromadb>=0.5.5